+++ tests/mocks/__init__.py
@@ -1,0 +1,8 @@
+class MockKeychain:
+
+    key = 'hieronymus_bosch'
+    spice = (
+        ".t`LueB[s;9>Qm1s&,bt3y8#bQ%@w='"
+        '<2lT>y}Al@[}/=9OFm""'
+        '#q2Isd?x|XF+'
+    )